Extract functions from bugwarrior.db.synchronize
This is just a refactoring that creates three new functions from parts of
synchronize().

The three functions in question are `create_task`, `modify_task` and
`close_task` and they encapsulate creating a new task, updating a task
or closing a task respectively.

Having separate functions not only makes synchronize a bit smaller it
also makes it easier to test (see included unit tests).